# apps/parishbooks-ui/openapi.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/session":{"get":{"operationId":"getCurrentSession","parameters":[],"responses":{"200":{"description":"Current user session"},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"pb_session_token":[]}],"summary":"Get the current session","tags":["auth"]}},"/api/v1/auth/sign-in":{"post":{"operationId":"signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"Sets pb_access_token, pb_refresh_token, and pb_session_token cookies when email is verified","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"429":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Sign in with email and password","tags":["auth"]}},"/api/v1/auth/sign-up":{"post":{"operationId":"signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"429":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Create an account and start email verification","tags":["auth"]}},"/api/v1/auth/sign-out":{"post":{"operationId":"signOut","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"pb_session_token":[]},{"pb_refresh_token":[]},{"pb_access_token":[]}],"summary":"Sign out and clear auth cookies","tags":["auth"]}},"/api/v1/auth/email-otp/send":{"post":{"operationId":"sendEmailOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendEmailOtpDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"429":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Send an email OTP","tags":["auth"]}},"/api/v1/auth/email-otp/verify":{"post":{"operationId":"verifyEmailOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailOtpDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"429":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Verify email with OTP and establish session cookies","tags":["auth"]}},"/api/v1/auth/forgot-password":{"post":{"operationId":"forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"429":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Request a password reset OTP","tags":["auth"]}},"/api/v1/auth/reset-password":{"post":{"operationId":"resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponseDto"}}}},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"429":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"summary":"Reset password with email OTP","tags":["auth"]}},"/api/v1/organization":{"get":{"operationId":"listOrganizations","parameters":[],"responses":{"200":{"description":"Organizations the current user belongs to"},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"pb_session_token":[]}],"summary":"List organizations for the current user","tags":["organization"]},"post":{"operationId":"createOrganization","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrganizationDto"}}}},"responses":{"201":{"description":"Organization profile created"},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"pb_session_token":[]}],"summary":"Create an organization and profile","tags":["organization"]}},"/api/v1/organization/{organizationId}/profile":{"get":{"operationId":"getOrganizationProfile","parameters":[{"name":"organizationId","required":true,"in":"path","description":"Organization id","schema":{"type":"string"}}],"responses":{"200":{"description":"Organization profile"},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"pb_session_token":[]}],"summary":"Get an organization profile","tags":["organization"]}},"/api/v1/organization/{organizationId}":{"get":{"operationId":"getOrganization","parameters":[{"name":"organizationId","required":true,"in":"path","description":"Organization id","schema":{"type":"string"}}],"responses":{"200":{"description":"Organization with profile"},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"404":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"pb_session_token":[]}],"summary":"Get an organization with its profile","tags":["organization"]}},"/api/v1/organization/activate":{"post":{"operationId":"setActiveOrganization","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SetActiveOrganizationDto"}}}},"responses":{"200":{"description":"Active organization updated; pb_access_token reminted with the new organizationId"},"400":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"pb_session_token":[]}],"summary":"Set the active organization for the current session","tags":["organization"]}},"/api/health/live":{"get":{"operationId":"healthLive","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["ok","degraded"],"example":"ok"},"info":{"type":"object","example":{"database":{"status":"up","responseTime":12}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up","responseTime":12}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["error","shutting_down"],"example":"error"},"info":{"type":"object","example":{"database":{"status":"up","responseTime":12}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect","responseTime":3005}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up","responseTime":12},"redis":{"status":"down","message":"Could not connect","responseTime":3005}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true}}}}}}}},"summary":"Liveness probe","tags":["health"]}},"/api/health/ready":{"get":{"operationId":"healthReady","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["ok","degraded"],"example":"ok"},"info":{"type":"object","example":{"database":{"status":"up","responseTime":12}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up","responseTime":12}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["error","shutting_down"],"example":"error"},"info":{"type":"object","example":{"database":{"status":"up","responseTime":12}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect","responseTime":3005}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up","responseTime":12},"redis":{"status":"down","message":"Could not connect","responseTime":3005}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true}}}}}}}},"summary":"Readiness probe","tags":["health"]}},"/api/health":{"get":{"operationId":"healthCheck","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["ok","degraded"],"example":"ok"},"info":{"type":"object","example":{"database":{"status":"up","responseTime":12}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up","responseTime":12}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","enum":["error","shutting_down"],"example":"error"},"info":{"type":"object","example":{"database":{"status":"up","responseTime":12}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect","responseTime":3005}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up","responseTime":12},"redis":{"status":"down","message":"Could not connect","responseTime":3005}},"additionalProperties":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["up","degraded","down"]},"responseTime":{"type":"number","description":"Time the health indicator took to respond, in ms"}},"additionalProperties":true}}}}}}}},"summary":"Combined health check","tags":["health"]}},"/api/v1/auth/refresh":{"post":{"operationId":"refreshAccessToken","parameters":[],"responses":{"200":{"description":"Sets a fresh pb_access_token cookie","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessResponseDto"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}},"429":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"security":[{"pb_session_token":[]},{"pb_refresh_token":[]}],"summary":"Remint pb_access_token from a valid session cookie","tags":["auth"]}}},"info":{"title":"ParishBooks API","description":"ParishBooks service OpenAPI document for generated clients (hey-api).","version":"1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"pb_access_token":{"type":"apiKey","in":"cookie","name":"pb_access_token"},"pb_refresh_token":{"type":"apiKey","in":"cookie","name":"pb_refresh_token"},"pb_session_token":{"type":"apiKey","in":"cookie","name":"pb_session_token"}},"schemas":{"ErrorResponseDto":{"type":"object","properties":{"statusCode":{"type":"number","example":400},"message":{"oneOf":[{"type":"string"},{"type":"array","items":{"type":"string"}}],"example":"Bad Request"},"error":{"type":"string","example":"Bad Request"},"timestamp":{"type":"string","example":"2026-10-05T04:00:00.000Z"},"path":{"type":"string","example":"/api/v1/auth/sign-in"},"correlationId":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"}},"required":["statusCode","message","timestamp","path"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string","example":"admin@parishbooks.com"},"password":{"type":"string","minLength":8,"maxLength":128,"example":"password123"},"rememberMe":{"type":"boolean","default":true}},"required":["email","password"]},"RedirectTo":{"type":"string","enum":["dashboard","email-verification","password-reset","org-setup","sign-in"]},"SignInResponseDto":{"type":"object","properties":{"redirectTo":{"allOf":[{"$ref":"#/components/schemas/RedirectTo"}]}},"required":["redirectTo"]},"SignUpDto":{"type":"object","properties":{"name":{"type":"string","example":"Jane Doe","minLength":1,"maxLength":100},"email":{"type":"string","example":"jane@parishbooks.com"},"password":{"type":"string","minLength":8,"maxLength":128,"example":"password123"}},"required":["name","email","password"]},"SuccessResponseDto":{"type":"object","properties":{"success":{"type":"boolean","example":true},"redirectTo":{"allOf":[{"$ref":"#/components/schemas/RedirectTo"}]}},"required":["success"]},"SendEmailOtpDto":{"type":"object","properties":{"email":{"type":"string","example":"jane@parishbooks.com"},"type":{"type":"string","enum":["email-verification","sign-in","forget-password"],"default":"email-verification"}},"required":["email"]},"VerifyEmailOtpDto":{"type":"object","properties":{"email":{"type":"string","example":"jane@parishbooks.com"},"otp":{"type":"string","minLength":6,"maxLength":6,"example":"123456"}},"required":["email","otp"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"jane@parishbooks.com"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"email":{"type":"string","example":"jane@parishbooks.com"},"otp":{"type":"string","minLength":6,"maxLength":6,"example":"123456"},"password":{"type":"string","minLength":8,"maxLength":128,"example":"newpassword123"}},"required":["email","otp","password"]},"OrganizationCountry":{"type":"string","enum":["IN","US"]},"OrganizationCurrency":{"type":"string","enum":["INR","USD"]},"CreateOrganizationDto":{"type":"object","properties":{"name":{"type":"string","example":"St. Mary Parish"},"slug":{"type":"string","example":"st-mary-parish"},"timezone":{"type":"string","example":"Asia/Kolkata"},"country":{"example":"IN","allOf":[{"$ref":"#/components/schemas/OrganizationCountry"}]},"currency":{"example":"INR","allOf":[{"$ref":"#/components/schemas/OrganizationCurrency"}]}},"required":["name","slug","timezone","country","currency"]},"SetActiveOrganizationDto":{"type":"object","properties":{"organizationId":{"type":"object","nullable":true,"example":"org_abc123","description":"Organization id to activate, or null to unset"},"organizationSlug":{"type":"string","example":"st-mary-parish","description":"Organization slug to activate when organizationId is omitted"}}}}}}